template-service-broker: (interesting) add cmd/template-service-broker command
